requirements.txt
aniso8601==10.0.0
antiorm==1.2.1
attrs==25.1.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
db==0.1.1
filelock==3.15.4
Flask==2.0.1
Flask-Cors==4.0.1
flask-restx==1.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PySocks==1.7.1
python-dotenv==0.19.0
pytz==2025.1
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rpds-py==0.23.1
setuptools==70.3.0
SQLAlchemy==1.4.27
tldextract==5.1.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.0.1
wheel==0.43.0
zope.event==5.0
zope.interface==6.4.post2
